+++ lib/InvertedClipper.dart
@@ -1,22 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:collection/collection.dart';
 import 'HoleArea.dart';
+import 'WidgetData.dart';
 
 class InvertedClipper extends CustomClipper<Path> {
-  final List<GlobalKey> keys;
-  double padding;
+  final List<WidgetData> widgetsData;
   Function deepEq = const DeepCollectionEquality().equals;
   List<HoleArea> areas = [];
 
-  InvertedClipper({this.keys, this.padding = 4}) {
-    if (keys.isNotEmpty) {
-      keys.forEach((key) {
-        if (key == null) {
-          throw new Exception("GlobalKey is null!");
-        } else if (key.currentWidget == null) {
-          throw new Exception("GlobalKey is not assigned to a Widget!");
-        } else {
-          areas.add(getHoleArea(key));
+  InvertedClipper({this.widgetsData}) {
+    if (widgetsData.isNotEmpty) {
+      widgetsData.forEach((WidgetData widgetData) {
+        if (widgetData.isEnabled) {
+          final GlobalKey key = widgetData.key;
+          if (key == null) {
+            throw new Exception("GlobalKey is null!");
+          } else if (key.currentWidget == null) {
+            throw new Exception("GlobalKey is not assigned to a Widget!");
+          } else {
+            areas.add(getHoleArea(key: key,shape: widgetData.shape,padding: widgetData.padding));
+          }
         }
       });
     }
@@ -26,8 +29,27 @@ class InvertedClipper extends CustomClipper<Path> {
   Path getClip(Size size) {
     Path path = Path();
     areas.forEach((HoleArea area) {
-      path.addOval(Rect.fromLTWH(area.x - (padding / 2), area.y - padding / 2,
-          area.width + padding, area.height + padding));
+      switch (area.shape) {
+        case WidgetShape.Oval: {
+          path.addOval(Rect.fromLTWH(area.x - (area.padding / 2), area.y - area.padding / 2,
+              area.width + area.padding, area.height + area.padding));
+        }
+        break;
+        case WidgetShape.Rect: {
+          path.addRect(Rect.fromLTWH(area.x - (area.padding / 2), area.y - area.padding / 2,
+              area.width + area.padding, area.height + area.padding));
+        }
+        break;
+        case WidgetShape.RRect: {
+          path.addRRect(RRect.fromRectAndCorners(Rect.fromLTWH(area.x - (area.padding / 2), area.y - area.padding / 2,
+              area.width + area.padding, area.height + area.padding),
+              topLeft: Radius.circular(5.0),
+              topRight: Radius.circular(5.0),
+              bottomLeft: Radius.circular(5.0),
+              bottomRight: Radius.circular(5.0)));
+        }
+        break;
+      }
     });
     return path
       ..addRect(Rect.fromLTWH(0.0, 0.0, size.width, size.height))
@@ -36,6 +58,6 @@ class InvertedClipper extends CustomClipper<Path> {
 
   @override
   bool shouldReclip(InvertedClipper oldClipper) {
-    return oldClipper.padding != padding || (!deepEq(oldClipper.areas, areas));
+    return !deepEq(oldClipper.areas, areas);
   }
 }