+++ lib/TutorialOverlayUtil.dart
@@ -1,8 +1,11 @@
 import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'OverlayPainter.dart';
+
 import 'InvertedClipper.dart';
+import 'OverlayPainter.dart';
+import 'WidgetData.dart';
 
 void showOverlayEntry(BuildContext context, OverlayEntry entry) {
   Overlay.of(context).insert(entry);
@@ -21,19 +24,25 @@ Future waitForFrameToEnd() async {
 }
 
 OverlayEntry createTutorialOverlay(
-    {@required BuildContext context,
-    List<GlobalKey> enabledKeys = const [],
-    List<GlobalKey> disabledKeys = const [],
+    {List<WidgetData> widgetsData = const [],
+    Function onTap,
+    Color bgColor,
     Widget description}) {
   return OverlayEntry(
-      builder: (BuildContext context1) => FutureBuilder(
+      builder: (BuildContext context) => FutureBuilder(
           future: waitForFrameToEnd(),
-          builder: (BuildContext context, AsyncSnapshot snapshot) => ClipPath(
-              clipper: InvertedClipper(keys: enabledKeys),
-              child: CustomPaint(
-                child: Container(
-                  child: description,
-                ),
-                painter: OverlayPainter(context: context, keys: disabledKeys),
-              ))));
+          builder: (BuildContext context, AsyncSnapshot snapshot) =>
+              GestureDetector(
+                  onTap: onTap,
+                  child: ClipPath(
+                      clipper: InvertedClipper(widgetsData: widgetsData),
+                      child: CustomPaint(
+                        child: Container(
+                          child: description,
+                        ),
+                        painter: OverlayPainter(
+                            bgColor: bgColor,
+                            context: context,
+                            widgetsData: widgetsData),
+                      )))));
 }