+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1] - 27/04/19
+
+* included onTap event
+* removed context parameter requirement from createTutorialOverlay()
+* breaking changes to API - now uses WidgetData class 
+                            instead of list of Global Keys.
+* added more shapes - (formerly supported only Oval) now also supports 
+                      Rect and RRect.
+* added bgColor property to createTutorialOverlay() to allow changing
+  default background color                                          
+
 ## [0.0.1] - 27/04/19
 
 * Initial release